requirements.txt
pyqt6
pillow